# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hypermart grocery store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ProductVariant {
  id: number;
  label: string;
  price: number;
}

export interface Product {
  id: number;
  name: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  discountPercent?: number | null;
  imageUrl: string;
  categoryId: number;
  categoryName: string;
  rating: number;
  reviewCount: number;
  sold: number;
  /** @nullable */
  badge?: string | null;
  isPromo?: boolean;
  variants?: ProductVariant[];
}

export interface ProductDetail {
  id: number;
  name: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  discountPercent?: number | null;
  imageUrl: string;
  images?: string[];
  categoryId: number;
  categoryName: string;
  rating: number;
  reviewCount: number;
  sold: number;
  /** @nullable */
  badge?: string | null;
  isPromo?: boolean;
  description: string;
  shelfLife: string;
  deliveryInfo: string;
  variants: ProductVariant[];
  relatedProducts?: Product[];
}

export interface Category {
  id: number;
  name: string;
  productCount: number;
  imageUrl: string;
}

export interface CartItem {
  productId: number;
  name: string;
  price: number;
  imageUrl: string;
  quantity: number;
  /** @nullable */
  variantLabel?: string | null;
}

export interface Cart {
  items: CartItem[];
  totalItems: number;
  totalPrice: number;
}

export interface CartItemInput {
  productId: number;
  quantity: number;
  /** @nullable */
  variantId?: number | null;
}

export interface CartItemUpdate {
  quantity: number;
}

export type ListProductsParams = {
/**
 * @nullable
 */
categoryId?: number | null;
/**
 * @nullable
 */
sort?: ListProductsSort;
/**
 * @nullable
 */
search?: string | null;
/**
 * @nullable
 */
limit?: number | null;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort] | null;


export const ListProductsSort = {
  popular: 'popular',
  terbaru: 'terbaru',
  teratas: 'teratas',
  harga: 'harga',
} as const;