# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hypermart grocery store API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductVariant } from './productVariant';

export interface Product {
  id: number;
  name: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  discountPercent?: number | null;
  imageUrl: string;
  categoryId: number;
  categoryName: string;
  rating: number;
  reviewCount: number;
  sold: number;
  /** @nullable */
  badge?: string | null;
  isPromo?: boolean;
  variants?: ProductVariant[];
}